/auth/logout endpoint was added, transfer from one account to another with transaction was added

+++ type.go
@@ -34,8 +34,8 @@ type LoginRequest struct {
 }
 
 type TransferRequest struct {
-	ToAccount int `json:"toAccount"`
-	Amount    int `json:"amount"`
+	ToAccount int   `json:"toAccount"`
+	Amount    int64 `json:"amount"`
 }
 
 type CreateAccountRequest struct {
